# swagger.yaml
definitions:
  storage.User:
    properties:
      age:
        type: integer
      created:
        type: string
      email:
        type: string
      firstname:
        type: string
      id:
        type: string
      lastname:
        type: string
    required:
    - age
    - email
    - firstname
    - lastname
    type: object
  storage.UserInput:
    properties:
      age:
        type: integer
      email:
        type: string
      firstname:
        type: string
      lastname:
        type: string
    required:
    - age
    - email
    - firstname
    - lastname
    type: object
info:
  contact: {}
paths:
  /:
    get:
      consumes:
      - application/json
      description: get users
      produces:
      - application/json
      responses: {}
      summary: List users
      tags:
      - users
  /create:
    post:
      consumes:
      - application/json
      description: post user
      parameters:
      - description: User data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/storage.UserInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/storage.User'
            type: array
      summary: new user
      tags:
      - users
  /edit/{id}:
    patch:
      consumes:
      - application/json
      description: updated user ID
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: User input
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/storage.UserInput'
      produces:
      - application/json
      responses: {}
      summary: updated user ID
      tags:
      - users
  /user/{id}:
    get:
      consumes:
      - application/json
      description: get string by ID
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/storage.User'
            type: array
      summary: Show a user
      tags:
      - users
swagger: "2.0"